requirements.txt
numpy==1.24.3
pyserial==3.5